requirements.txt
altair==5.5.0
amqp==5.3.1
anyio==4.12.1
appnope==0.1.4
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
billiard==4.2.1
blinker==1.9.0
CacheControl==0.14.4
cachetools==5.5.2
celery==5.5.3
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.3
cryptography==45.0.5
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
dj-rest-auth==7.0.1
Django==5.1.6
django-allauth==65.10.0
django-cors-headers==4.7.0
django-environ==0.12.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
ecdsa==0.19.1
et_xmlfile==2.0.0
executing==2.2.0
firebase-admin==6.9.0
flatbuffers==25.2.10
fonttools==4.61.1
geographiclib==2.1
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-core==2.5.0
google-cloud-firestore==2.25.0
google-cloud-storage==3.4.1
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.70.0
grpcio==1.78.0
grpcio-status==1.78.0
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
h3==4.3.1
holidays==0.87
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
inflection==0.5.1
ipykernel==6.29.5
ipython==8.32.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.9
kombu==5.5.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.8
matplotlib-inline==0.1.7
mdurl==0.1.2
msgpack==1.1.2
narwhals==1.37.0
nest-asyncio==1.6.0
numpy==2.2.5
oauthlib==3.3.1
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.6
plotly==6.5.0
prompt_toolkit==3.0.50
proto-plus==1.26.1
protobuf==6.33.6
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycountry==24.6.1
pycparser==2.22
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.9.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.2.1
redis==6.4.0
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.24.0
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
smmap==5.0.2
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
stack-data==0.6.3
streamlit==1.55.0
streamlit-option-menu==0.4.0
stripe==15.0.1
tenacity==9.1.2
threadpoolctl==3.6.0
timezonefinder==8.0.0
toml==0.10.2
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
vine==5.1.0
watchdog==6.0.0
wcwidth==0.2.13
whitenoise==6.9.0
xgboost==2.1.4
